REFACTOR: separate DockerExecContainer tests into its own class
